+++ CHANGELOG.md
@@ -5,8 +5,14 @@ title: Changelog
 
 ### 0.0.11 (Unreleased)
 
+`uneventful`
 - Refactored internal context management to use less memory (and fewer objects) per signal, and to reduce the amount of pointer indirection on some common code paths.
+
+`uneventful/signals`
 - Fixed an issue where signals polling external data using {@link uneventful.recalcWhen}() could become stale unless observed by a rule.
+
+`uneventful/utils`
+- Added {@link uneventful/utils.call `call()`} function as an IIFE replacement utility
 
 ### 0.0.10