+++ CHANGELOG.md
@@ -3,7 +3,7 @@ title: Changelog
 ---
 # Changelog
 
-### 0.0.11 (Unreleased)
+### 0.0.11
 
 `uneventful`
 - Refactored internal context management to use less memory (and fewer objects) per signal, and to reduce the amount of pointer indirection on some common code paths.